Adds a view helper to build navigation menus.
This view helper can build arbitrary menus as nested lists. It can definately be
improved, check comments in the source. So far, it simply links to the index
action of the index controller of each specified module. It does use
translations, it expects each module to define a translation string named
{modulename}_modulename.

git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@943 0a49f362-4213-0410-86e3-e96cb6c4fbae