Adds constructing Opus_Model_Documents by type name.
Opus_Model_Document now creates it's own Opus_Document_Builder instead of taking
one as a parameter. When passing a document type name, a new document of that
type is contructed. When passing an id, the type is read from the corresponding
database entry.

Also adds a couple of necessary mock up _fetch methods in Opus_Model_Document
and a function to get all documents a person is associated to by a certain role
to Opus_Model_Person.

git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@1054 0a49f362-4213-0410-86e3-e96cb6c4fbae