Adds method to convert Opus_Models to an array.
This adds a toArray() method to Opus_Models. The method returns a nested array representing the model. This is mainly
aimed at being used in views, since Opus_Models contain some database details that should not leak, i.e. an Opus_Model
should never be used directly in a view.

git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@1529 0a49f362-4213-0410-86e3-e96cb6c4fbae