Adds SMTP honeypot for tests.
Execute 'sudo php smtp_honeypot.php' (you'll need to be root if you want it to listen on default port 25). This will
start a pseudo SMTP server, dumping all communication to the screen and only simulating sending email. With a bit more
work, this could be used for unit testing the mail classes that require an actual mailserver.

git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@1925 0a49f362-4213-0410-86e3-e96cb6c4fbae